+++ game/GreenMushroom.cpp
@@ -41,7 +41,7 @@ void GreenMushroom::Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects)
 		FilterCollision(coEvents, coEventsResult, nx, ny);
 	}
 
-	for (UINT i = 0; i < coEvents.size(); i++) delete coEvents[i];
+	for (size_t i = 0; i < coEvents.size(); i++) delete coEvents[i];
 }
 
 void GreenMushroom::Render()
